[llava][15/N] Extract out text decoder runner
Last PR #4556 refactored run_model_step() so that it is suitable to be
extracted out as a separate class.

This new `TextDecoderRunner` provides 2 APIs:

* step(tokens, start_pos)
This API takes one or more tokens with start_pos and feed them into Module. Return a tensor of logits.

* logits_to_token(logits)
This API samples the result and returns a token. We don't expect this
logic to change across different runners.

ghstack-source-id: 2628a3e88dbbd11bddc78062bebe361bbd2c529e
Pull Request resolved: https://github.com/pytorch/executorch/pull/4567